Use local repo interface in member usecase
The member usecase profile-interaction defines a local interface
memberRepo. It represents the contract that is needed to fullfill the
requirements of the usecase.
The io/dummy/ package contains a dummy implementation of this local
interface.
The main.go now assembles the app, initializes the needed usescases and
injects the concerete technical componentens (in this case a repo that
provides a GetMemberByID method.