Move tests out test package and add build tags
When tests are in a seperate test package the tooling (go test -cover,
goconvey, travis) does not work properly.
So we keep the tests in the same package as the component under test.

Additionaly we're using build tags to classify different test types.
Simply add a comment at the top level of a test file:
// +build learning
// +build io
Currently there are three types of tests.
- learning tests are simple playgrounds for learning 3rd Party APIs.
- io tests tries to interact with the outside world. So they might need
a valid API token to access external services.

Tests with build tags are NOT executed by the CI by default.
So if you run `go test ./...` (like CI does) only those tests are
executed that does not have a build tag.

If you want to execute the special tests (like learning or io) you have
to use the `-tags` flag. For example `go test -tags="learning io"
./...`.

Make sure, that the default tests (without a buil tag) does not use
external services and are short running, so that you can execute them
continuously during coding".